ENH: create AcqBaseWindow in common.afc_acq_core and use for FOFB acquisition windows
